menu refactoring
