+++ client/src/components/layouts/Layout/Header/Menu/MenuNavLink.tsx
@@ -9,13 +9,7 @@ interface Props {
 
 const MenuNavLink: FC<Props> = ({ link, title }) => {
   return (
-    <NavLink
-      className="dropdown-item"
-      to={link}
-      onClick={() => {
-        document.body.click();
-      }}
-    >
+    <NavLink className="dropdown-item" to={link} onClick={() => document.body.click()}>
       <span>{title}</span>
     </NavLink>
   );
