--- client/src/components/layouts/Layout/Header/styles.scss
@@ -1,166 +1,0 @@
-@import 'src/variables.scss';
-
-.navigation-header {
-  border-bottom: 1px solid $light-gray-color;
-  background: $dark-gray-color;
-  color: $white-color;
-  .header-logo {
-    display: flex;
-    justify-content: center;
-    img {
-      height: 20px;
-    }
-  }
-}
-
-.header-nav-dropdown {
-  flex-basis: 0;
-
-  a#headerNavDropdown {
-    display: flex;
-    justify-content: start !important;
-  }
-  .dropdown-toggle {
-    margin: 0 .75em !important;
-    width: 20px;
-    height: 20px;
-    display: inline-block;
-    background: no-repeat center url('/assets/images/hamburger.svg');
-    color: red;
-
-    &::after {
-      display: none !important;
-    }
-  }
-
-  &.show .dropdown-toggle {
-    background-image: url('/assets/images/close.svg');
-    width: 20px;
-    height: 20px;
-  }
-
-  .nav-link {
-    padding: 0;
-    color: white !important;
-  }
-
-  .dropdown-menu {
-    position: fixed;
-    top: 36px !important;
-    margin: 0 !important;
-    padding: 0 !important;
-    border: 0 !important;
-    background: $dark-gray-color !important;
-
-    &.show {
-      width: 100vw !important;
-      margin: 0 !important;
-      border-radius: 0 !important;
-    }
-  }
-
-  .nav-dropdown-menu {
-
-    .dropdown-menu-user-name {
-      color: $white-color;
-      font-weight: 400;
-      max-width: 260px;
-    }
-
-    .dropdown-menu-user-picture {
-      width: 36px;
-      height: 36px;
-    }
-
-    .dropdown-item {
-      font-size: 14px;
-      color: $white-color;
-      padding: 0 0 1em 0 !important;
-      letter-spacing: 0.05rem;
-      &.active {
-        background: none !important;
-        font-weight: 700;
-      }
-
-      &:hover {
-        background-color: $barely-gray-color;
-      }
-    }
-
-    .dropdown-menu-social-media {
-      a {
-        background-repeat: no-repeat;
-        background-size: contain;
-        width: 36px;
-        height: 36px;
-      }
-
-      .twitter {
-        background-image: url('/assets/images/twitter.webp');
-      }
-
-      .instagram {
-        background-image: url('/assets/images/instagram.webp');
-      }
-
-      .facebook {
-        background-image: url('/assets/images/facebook.webp');
-      }
-    }
-
-    .dropdown-menu-copyright, .dropdown-menu-privacy {
-      color: $gray-color;
-    }
-
-    .dropdown-menu-privacy {
-      a {
-        color: $gray-color;
-        text-decoration: none;
-        font-weight: normal;
-        letter-spacing: 0.025rem;
-      }
-    }
-  }
-}
-
-.desktop-navbar {
-  display: none !important;
-  .navbar-nav {
-    a {
-      color: $barely-gray-color !important;
-      font-size: 11px;
-      font-weight: 500;
-      letter-spacing: .025rem;
-    }
-  }
-}
-
-.header-login-profile-icon {
-  margin-right: .75em !important;
-  text-align: right;
-  flex-basis: 0;
-  img {
-    height: 20px;
-  }
-}
-
-@media (min-width: $medium-weight) {
-  .navigation-header {
-    .header-logo {
-      flex-basis: 0;
-      justify-content: start;
-    }
-  }
-  .header-login-profile-icon {
-    margin-right: 0 !important;
-  }
-  .header-nav-dropdown {
-    display: none !important;
-  }
-  .desktop-navbar {
-    display: block !important;
-    .navbar-nav {
-      justify-content: center;
-    }
-  }
-}
